pushed for checking or errors encountered

+++ reports.h
@@ -4,7 +4,7 @@
 void display_account();
 void month_report();
 void box_for_display();
-void return_name(long int accno, char *name_out);
-void return_address(long int accno, char *address_out);
+void return_name( int accno, char *name_out);
+void return_address( int accno, char *address_out);
 
 #endif
